docs, Button: remove a number of selection demo variations
Now that the configurable demo supports configuring an icon to use
in the content slot, there is no need for so many specific variations
to be shown. Remove the demos for

* Progressive, with icon
* Destructive, primary, with icon
* Progressive, primary, icon only
* Destructive, icon only

Also, fix the sample code that is used for the "Default, quiet, with icon"
example, which had been using the "Destructive, primary, with icon" code
to show instead.

Change-Id: Iefcde61a84e6273840c0a84dba6a017537fc0672