docs, tokens: add generic CdxDocsTokenDemo for demonstrations
Work for most of the existing demos, which are replaced. Affected
demo components:
* CdxDocsAnimationDemo
* CdxDocsBorderDemo
* CdxDocsBoxShadowDemo
* CdxDocsColorDemo
* CdxDocsCursorDemo
* CdxDocsOpacityDemo
* CdxDocsPaddingDemo
* CdxDocsPositionDemo

Token demos that formerly used one of those were updated to
instead use CdxDocsTokenDemo and specify the specific CSS
class that should be used for their specific extra styles.

Styles for these demos were moved as-is to CdxDocsTokenDemo,
with two small exceptions:
* the animation __bar class and transition were renamed to __inner

* the outline tokens now use the new class `cdx-docs-outline-demo`
rather than `cdx-docs-box-shadow-demo`, though it has the same
styles

The style duplication in CdxDocsTokenDemo will be cleaned up in
a follow-up commit.

Change-Id: I93d2e7aa71ce7dc7512fbb8142139db775cfd8a6