build: Add separate entry point for components in development
Components in the src/components-wip directory are in the repo,
but are not yet ready to be included in the library. They are not added
to lib.ts, and so are not included in releases of the library.

However, we do want the demo site to have access to work-in-progress
components, so that these components can have demos. To make that
possible, add a second entry point file, lib-wip.ts, which
exports all components from src/components-wip/index.ts
in addition to the ones from lib.ts. The demo site then aliases
'@wikimedia/codex' to lib-wip.ts.

To make the demos not fail TypeScript, configure TypeScript in the demo
package to look at lib-wip.ts rather than lib.ts.

Also use an eslint rule to disallow imports from paths that involve the
"components-wip/" directory. This allows internal imports within
that directory, but disallows code outside that directory from importing
anything in that directory.

This allows new components to first be added in the components-wip
directory, then undergo further development before they're moved to the
components directory and are included in the library.

Bug: T313767
Change-Id: Ib91819a91e3f53e39970622ea6950c13f2d87bc8