MenuItem, TypeaheadSearch, styles: Fix link style overrides
The `.cdx-menu-item__content` element could be an anchor tag, in
which case we need to override color and text-decoration styles
that might be set on all `a` elements (for example, in MediaWiki).

This patch specifically applies color and text-decoration styles
to this element to prevent unwanted styles from polluting this
component.

Bug: T309727
Change-Id: I14762abb020c90b26a8a471f81b0ce8649b32d98