TypeaheadSearch, styles: Fix auto-expand distance
To keep the cursor and input text in a static position when the
input is auto-expanded on focus, a margin-left is placed on the
root element of TypeaheadSearch that equals the amount the input
grows when expanded. This is calculated by taking the padding-left
of the input when expanded and subtracting the padding-left of the
input when not expanded.

The padding-left of the input when not expanded was recently
changed (reduced), so we need to change (increase) the margin-left
accordingly. This patch does that, and generally changes the way
this value is calculated to match how we're calculating the
input padding-left in both scenarios.

Bug: T309733
Change-Id: I1a7cd51b4d954d00033cf0b628efdda0aceb63fe