TypeaheadSearch: Remove hover effect from button
Remove the styling that makes the button only appear on hover. Consumers
that want this effect can create it in their own CSS.

Also make the buttonLabel prop optional, and hide the button if the
label isn't set, just like in SearchInput. Add an example of this to the
docs.

Minor cleanup:
- Add a class to the TextInput component so that we can target that in
  CSS, rather than using .cdx-text-input
- Replace border-radius rules with .element-with-menu-expanded() mixin
- Remove unused position: relative; rule. There was a left: rule here
  once, but it was removed a while ago.
- Clarify that autoExpandWidth does nothing without showThumbnail

Bug: T309722
Change-Id: I0c8f08f534d358f1196222b12660fd2dc6d38cb0