MenuItem: Remove highlighted and active styles on mouseleave
MenuItems are highlighted on mouseenter (aka hover), and if you
drag the mouse from one MenuItem to another, the new MenuItem
will be highlighted. However, if you drag your mouse off of the
Menu altogether, the last MenuItem will remain highlighted.
Similarly, mouseleave during mousedown will result in a
still-active Menuitem.

This patch removes the highlighted item on mouseleave, and
changes the criteria for setting the `--active` class on
the MenuItem component to include that the item must also
be highlighted to be shown as active. This way, the Menu
component will continue to consider a MenuItem active on
mouseleave, but the active styles will disappear. Then,
on subsequent mouseenter, the active styles will reappear.

Bug: T308170
Change-Id: I8dc565ec08929ced9f33de5dcc191da83608b8b7