MenuItem: Align icon and thumbnail to top with description
This patch sets the alignment of the icon/thumbnail and text to
the top, rather than center, when a description is present. It
also makes the icon size relative so icon will scale with
font size, leading to a nicer look when font size changes.

Thumbnail size was not made relative due to TypeaheadSearch
styles that are calculated based on the pixel size of the
thumbnail.

Additionally, the line-height of MenuItem content was set to
1.375 (22/16), as specified by the design spec, which improves
the alignment of a MenuItem icon with the title.

Finally, the initial icon used in the configurable MenuItem
demo was changed to something more square, which looks a bit
more natural when the MenuItem is displayed out of the context
of a Menu. The shape of the old icon (tag) caused the
alignment to look a bit off, even though it's correct.

Bug: T306525
Change-Id: Ic25d7f7b90146c4c47da738262c20ef5e0b2a883