tokens: Fix quiet active background-color value
And also align with WikimediaUI Base CSS way of using `calc()` for
simpler value understanding.
This is the second time amending this value as the division wasn't
applied correctly last time.
Note, that we are currently not featuring an `rgba()` color in Figma
for those states, but discussion is set out on task to get on same
page for use case of RGBA vs HEX color notation.

Bug: T305402
Change-Id: I2409aac0f45f215380f3eb29dfb8717f5b901ecb