TextInput: Remove focus tracking state, replace with --has-value class
Remove the isFocused variable, and instead just apply a CSS class based
on whether the input is empty. To change the icon color both on focus
and on non-emptiness, use a comma selector in CSS instead of one class
that expresses both.

This change will help us temporarily disable the icon color transition
on focus in Vector, because doing this using native :focus instead of
through a CSS class makes the style change happen earlier.

Change-Id: If6991f74412e54e88c989b0f0e4b927281b9067c