tokens, styles: Use `transition-property-base` for ToggleSwitch focus
Using `transition-property-base` for ToggleSwitch's focus outline
instead of SFC token. Normally we would limit transition property
to specific CSS property, here `border-color` to avoid side-effects.
In this case it's acceptable to use base token to keep tokens amount
leaner and due to the fact that there are no side-effects to expect
on this specific selector.

Bug: T295711
Change-Id: I30280c72f9a0effa60066a2570b24bbe600de1d6