docs: Update CSS class names on tokens demos
Updating TokenDemo and TokensTable to provide a more structured
CSS class hierarchy, that will allow for specific token demos to
be treated differently. It bring's tokens demos closer to a modular
BEM structure.
With the current demos need of having to style
1. all tokens demos,
2. tokens demos of a specific token group and
3. specific tokens
differently together with the complexity of nested wrapper and inner
elements of a demo (e.g. Position demo), this approach is needed to
fulfill above.

Also
- introducing new `tokenCategory` prop and using `demoClass` only for
  static class
- expanding code comments
- fixing 'cursor.md' TokensTable call

Bug: T311082
Change-Id: Ic1ecd43cec0dd5435aa446482a043be4c44f7c50