Menu: Refine `active` binding of default slot
The `active` binding was previously based on whether the menu
item was also the menu's current active item. However, there
is a differentiation between the menu's current active item
and the current visually active item: an active item only has
active styles applied if it is also highlighted. This allows
the menu to continue tracking it as active in the case of
mousedown -> mouseleave, in case it is highlighted again.

This patch updates the logic of the `active` binding of the
default slot to check for both active and highlighted status
within the menu. As a result, uses of the default slot, e.g.
the search footer of TypeaheadSearch, will behave the same
way as menu items using the default slot content.

Bug: T308170
Change-Id: I1f125c51176f72ee90f59f64c1ec4a16d38d46ad