tokens: Add design-first breakpoints tokens
Implementing design-first `min-width` and `max-width` breakpoints.
Keeping the `width` based breakpoints as legacy tokens for
non-Codex products WikimediaUI Base.
Note that for `max-width` breakpoints, we currently rely on `calc()`
as it has full support in all browsers we support. And we need to
rely on pure CSS output as well for the Codex design tokens.

Also add Breakpoint tokens demo page.

In option tokens we group them separately from size and spacing tokens
for their single use case.

Decision record to the breakpoint token decision naming convention
updated on task description.

This builds on top of Figma Breakpoints v10 of 2022-08-03.

Bug: T303522
Change-Id: Iaf68d7e6b656a09faf5a7802bb829fa59bdec174