ButtonGroup: Initial implementation
ButtonGroup creates a series of buttons based on ButtonGroupItem
objects. When a button is clicked ButtonGroup emits a 'click' events
with the .value property of the clicked button.

Each button is shifted left 1px so that its left border is on top of the
previous button's right border, deduplicating the borders. We use
z-index trickery to make active and focused buttons appear above the
rest of the buttons, so that both of their borders are visible, and to
make disabled buttons appear below the other buttons so that both of
their neighbors' borders are visible.

Consecutive disabled buttons are styled to have a white border. All
internal borders use straight corners, only the external borders have
rounded corners.

Bug: T311756
Change-Id: I93cbdfadebe93c3f4ac3911c93161561b3fb9ff9