ToggleButtonGroup: Initial implementation
ToggleButtonGroup creates a series of toggle buttons based on ButtonData
objects, and aggregates their v-model values into a single value.
Whether only a single selection is allowed or multiple selections are
allowed is based on whether the initial value of v-model is an array or
not.

The styles for adjacent buttons are shared with ButtonGroup. In
addition, a white border is added between adjacent toggled-on buttons,
and toggled-on buttons appear above their neighbors so that their
borders are visible on both sides (but appear below active or focused
buttons).

Bug: T310768
Change-Id: If5b1698d5e4cb6c72ec90a7288bf443610c080bc