docs: Add step to releasing docs to document breaking changes
As part of the release process, breaking changes should be
briefly explained in the CHANGELOG to help users of Codex
understand what they should expect and what they can do to
handle the changes in their codebases.

Docs for other parts of the process, e.g. announcing that
breaking changes are coming, will be handled elsewhere.

Bug: T312167
Change-Id: I81f50a4694191190cd106b88d6aed77b5c876dc7