Button: Ignore whitespace-only text nodes when detecting icon-only-ness
If the contents of a button are something like:
    <cdx-icon :icon="foo" />
    {{ label }}

but `label` is an empty string, then the resulting VNode array is an
icon component node followed by a text node that contains a single
space. This whitespace-only text node confuses the icon-only detection
code and makes the button appear as not being icon-only, even though it
is.

To fix this, ignore whitespace-only text nodes.

Follow-Up: If190010ebb29c27b5e096d2c545f0649e6f16947
Change-Id: I56291197aaa905f73aa1795ef5063710a8f4757b