TypeaheadSearch: Open menu on new results, even if empty
When deciding whether to open the menu when the searchResults prop
changes, one of the criteria we were looking at is whether the array of
search results we received is empty. But that shouldn't matter: we
always display the search footer, so we should always open the menu when
results arrive.

What we should be checking for instead (and what the code comment
already incorrectly claimed we were checking for) is whether the search
input is empty.

Bug: T314089
Change-Id: I29103fbef55027e841912a3bc0d6e7dacf4eea5d