docs: Flag development components, hide them in release docs
Detect whether a component is under development by looking whether it's
in the src/components-wip directory. If it is, add a warning
banner at the top of its documentation page.

If the docs are built with `npm run doc-release`, don't list development
components in the sidebar at all. Building the docs the regular way
(with `npm run doc`) or previewing them with `npm run doc:dev` does
list them in the sidebar, but with a construction emoji symbol to
indicate work in progress status.

Even in the release build, the docs for development components can
still be accessed if you know the URL.

Bug: T313767
Change-Id: I705c084e4025e136f438a6dc2117fc7c877cd40d