ButtonGroup, ToggleButtonGroup: Straighten white lines between buttons
When two disabled buttons or two toggled-on buttons are adjacent to each
other, we show a white line between them (by giving the second button a
white left border). That white line was diagonally cut off at the ends
because the buttons also have top and bottom borders whose colors match
the background color. Remove these top/bottom borders so that the white
line is no longer cut off.

Changing the border-width here doesn't affect the appearance or
dimensions of the button, because the top/bottom borders were the same
color as the background, and because box-sizing: border-box; is set.

Bug: T310768
Change-Id: I0410314c7da72dd3771719369484a8a1f24a3095