ButtonGroup: Apply rounded corners to groups, not buttons
This way rounded corners look better when the buttons wrap to the next
line. Instead of the first and last buttons having rounded corners, the
buttons in the corners have rounded corners.

Without this, the bottom left corner of the leftmost button on the
bottom row is not rounded.

- Unset border-radius on individual buttons
- Apply border-radius to the group instead
- Set overflow: hidden; so that the group's border-radius causes the
  buttons to have rounded corners
- Set width: fit-content; so that the rightmost button still gets
  rounded corners in a group that doesn't wrap to the next line,
  even if the buttons don't fill all the available width

Stylelint warns on width: fit-content; because intrinsic widths aren't
fully supported on all browsers we target, but caniuse says we're fine.
None of the caveats at https://caniuse.com/intrinsic-width apply to
width: fit-content; and
https://caniuse.com/mdn-css_properties_width_fit-content is green for
all browsers we target.

Bug: T311756
Change-Id: I709f1d8664d348cc1ae8ce2fa24a36c04cad17fe