CopyTextButton: Use Clipboard API when available to copy code
The Clipboard API is only available in secure context, use that when
available. If not, continue to use the document.execCommand('copy')
approach to copy the content.

The Clipboard API was tested by adding my local domain under
"Insecure origins treated as secure" option in chrome://flags

Using the Clipboard API also fixes the issue with extra newline
getting copied in Chrome

Bug: T306573
Change-Id: I238217d800e66bc47cd29da54d6075b279bd0469