ButtonGroup: Use box-shadow instead of border between disabled buttons
There were issues with using border-left-color: white; to draw a white
line between two disabled buttons (or two toggled-on buttons). Mainly,
 when the second of two adjacent disabled buttons wraps to the next
line, the border is still drawn and the button appears misaligned with
the button above it.

Instead of using a border, use a box-shadow. This doesn't suffer from
this problem, doesn't affect the box model sizing of the buttons, and
also allows us to draw white lines between vertically adjacent disabled
buttons. We also don't have to use the + combinator anymore, because we
can just apply the box-shadow to all disabled buttons, and if the
adjacent button is a non-disabled button it'll have a higher z-index and
the box-shadow will be invisible below it.

Also apply the same box-shadow to toggled-on buttons. This causes minor
imperfections when a toggled-on button is adjacent to another kind of
button, because the neighboring button will have a lower z-index, not a
higher one. But overall this is still an improvement.

Bug: T311756
Bug: T310768
Change-Id: Ib5fe17628f8551c93157912caf1403bf5d108333