Lookup: When input is empty, clear pending state, and don't reopen menu
The "lookup with fetched results" demo had a bug where, if you typed a
character and then quickly deleted it while the component was still in
the pending state, the pending state would remain. This was in part
because the demo code just did nothing when the input value was empty,
rather than setting menuItems to an empty array.

This change fixes the demo code to set menuItems to an empty array
before returning early, but that initially revealed another bug: clearing
the input caused the menu to open with a no results message.

Fix both bugs by explicitly clearing the pending state when the input is
emptied. This also fixes the issue of the menu reopening with an empty
input, because we won't open the menu in response to a menuItems change
when we're not in the pending state.

Bug: T313416
Change-Id: Id7aee0e108302af9c3d2f263505359c8a89aea6e