docs, tokens: Show deprecated tag even if there is no token demo
The "deprecated" tag was inside a <div> that was only shown if there was
a token demo, but it wasn't related to that div at all. It was
absolutely positioned, but relative to that div's parent. Move it up one
level so that the "deprecated" tag is also shown for tokens that don't
have a demo.

Change-Id: I645b1a4e9115dcf03690da8decb28f2b3f9eb21b