docs, tokens: Exclude breakpoint tokens from the size token docs
The docs for width, min-width and max-width on the size tokens page list
the breakpoint tokens a second time, because the breakpoint tokens are
nested in these width token structures. The breakpoint tokens are
already displayed on the breakpoint page, we don't want to duplicate
them on the size page.

Add an `exclude-tokens` prop to the TokensTable component, and use it on
the size page to exclude the breakpoint tokens.

Change-Id: I8ef53b4a6712abb166e4295197b02070902d5125