docs: Reword alpha warning
We're planning to release 0.1.0 next, so remove the language about Codex
being in alpha and not ready for production use (since it's used in
production already!). Replace that warning with a warning about Codex
being incomplete and under active development.

Also change the "Versioning" section to no longer explain the -alpha and
-beta version numbers, and instead explain our practice of using 0.x.y
version numbers and incrementing `x` when a breaking change happens.

Change-Id: I684a28f239fc6da8ca3821739c4cb907001bb26c