ToggleButtonGroup: Combine focus shadow and white border shadow
Toggled-on buttons in a ToggleButtonGroup get a box-shadow that shows
an inverted white border on the left and top edges. When the button is
focused, that box-shadow is overridden by a focus style in ToggleButton
to show a blue focus shadow, so the white border shadow doesn't appear.
Fix this by combining the two shadows in ToggleButtonGroup, adding a
white outset shadow around all four sides of the button. This makes the
blue focus shadow and border more visible, because there's another white
shadow around it to separate it from the adjacent button (which may also
be blue if it's toggled on).

Although this new white outset shadow on focus appears on all four
sides, the basic white shadow that appears on unfocused toggled-on and
disabled buttons continues to only be on the left and top edges.

Bug: T310768
Change-Id: I4f3a8cb3bb51a703c02abbafe3e163d2f101bf6b