Dialog: Introduce Dialog component, useResizeObserver
Introduces Dialog as a work-in-progress component (not part of the
published distribution for now, but available in docs and local
development). Dialog can be used on its own, but may also be used
within Vue's <Teleport> component; this latter usage may be useful
inside of layouts where other absolutely positioned elements appear
after this component in the DOM.

Menu-based components (Select, Combobox, Lookup) receive some small
modifications to styling so that they can function properly when
used inside of a Dialog. When a Menu-bearing component is used
inside of the Dialog (which is already using absolute positioning),
the position of the outer element is set to "static" and the menu
is given a definite width equal to the current calculated width of
that element; this value is injected into the styles through Vue's
v-bind() CSS feature.

To support easy (and reactive) calculation of the dimensions of a
given element in a component, this patch also introduces a new
composable, useResizeObserver. This function is modeled after the
useIntersectionObserver composable, and acts as a simple wrapper
around the underlying API. Browsers that do not support
ResizeObserver will not benefit from this feature but components
using this composable will not otherwise break in such
environments.

Co-Authored-by: Roan Kattouw <roan@wikimedia.org>
Co-Authored-by: Anne Tomasevich <atomasevich@wikimedia.org>
Bug: T313773
Change-Id: Ib8161515f32f50467e218da0377580a92931e3fa