build: Exclude WIP code from test coverage threshold
Set a separate threshold of 0% for WIP code. This takes it out of the
global threshold computation.

We would like to enforce at least minimal test coverage for WIP
components, but unfortunately setting the threshold to a low nonzero
number (e.g. 1%) doesn't enforce that, because files that are never
imported are just completely ignored by Jest in our current
configuration. This may be fixable through another Jest config change,
but it's not part of this patch for now.

Bug: T316139
Change-Id: I85f450fe26add5cd02a8e1db8e5bf2d4736d20e6