docs: Hide direction switcher
The direction switcher does not play nicely with the new header
nav on certain screen sizes in the tablet range. Eventually,
the direction switcher will be moved out of the header and
into individual component demos. For now, we will hide the
direction switcher with CSS, so those of us who know about it
can un-hide it and use it for testing in the meantime.

Change-Id: I1e5bcb8c7cc7fb974f507a906806a5127edfba89