Menu: Emit 'load-more' event when close to the end
This patch also adjusts the component to not emit an event when an
options is selected that has the same label as the text that was typed
into the input.
That mostly prevents a bug that would occur when loading more results
and then selecting a option from later in the list. Without this
adjustment, the initial set of results would be reloaded from the
server, not including the original one.

Bug: T316603
Change-Id: Ie5fd6195547dda0711ea1cf000b2acdae97e7256