build: Point docs tsconfig to lib-wip.ts, not the .d.ts file
I don't know why I had pointed this to the type definition file instead
of to the entry point itself. This probably worked in CI because it
builds the type definitions first, but it's not a nice experience during
development.

Pointing the alias directly to lib-wip.ts should improve the developer
experience, although it may make running TypeScript on the docs site
slower, and may cause issues if the tsconfig of the docs site ever
diverges from that of the main library.

Change-Id: I6a81836a0d5b1ce093c0abc5db3f1a9584516153