docs: Restructure Codex docs introduction page 'index.md'
- Improve heading structure
- Remove first headline in the content
- Restructure further information
-- Put current version after introductory sentence
-- Add a tagline to the header instead
-- Adding additional paragraph who's behind Codex

Also, while at it, repeat the information about maintainers with links,
we recently introduced in the repository's README.

Removing on-page navigation (`aside`) in index.md page as it's more
distracting than helpful here.

On a technical level, I've boldly update `line-height` to coming
typographic token value of `1.6` instead of `1.7`. This makes the
introductory header more harmonious.
With the new `line-height`, we need to adjust the `code` styling too.
The reduced `padding` is inline with MediaWiki's `code` styling.

I've also removed an unnecessary `div` from the index page.

Bug: T318282
Change-Id: I3d60d551691adb791f3eb4e5298432c36c4b5e83