docs, tokens: Replace deprecated design tokens
Also remove custom border-color for `code` snippets. Let's go with
`border-color-subtle` for now and raise the need for an even subtler
one to the design team if needed.
Note, that `background-color-framed` and `background-color-interactive`
are carrying the same value, but former has been deprecated.

Lastly, replacing `opacity` static values with tokens.

Bug: T305411
Change-Id: I100679c0f1a416ff00aed72f5865f6a3c1b29f0a