tokens: Deprecate component `background-color-primary*` tokens
Both are not featured anymore in latest DS Color tokens v18 as of
2022-09-22.
https://www.figma.com/file/E6hxGwEOs9YFVF4fFRyqSv/Colors-Tokens---T296995?node-id=1650%3A28752

The `background-color-primary--hover` token is featured for backwards-
compatibility with WikimediaUI Base. Design has decided to not need
to rely on very sublte `background-color` state changes as text color
changes are sufficient in the few places currently used, for example
selected menu or VE tool state.

Change-Id: I0159827c49e14e9d503a75a12529f80088bd4eb7