tokens: Add design-first `text-decoration` tokens
Based on Typography DS Figma file, v11 as of 2022-09-22
https://www.figma.com/file/X8pKlndyPaqZg4I3GubQs6/Typography?node-id=886%3A5584

We are aware that the decision token names are not application
specific, but due to the fact that there has not been found a clear
umbrella term for e.g. all the different text decoration underline
styles, relying on the CSS property name is the best option for the
decision tokens for now. Especially given that it's a relatively rarely
used token in a wide mix of uses and the risk of confusion is low.

Bug: T318023
Change-Id: Ic724fdf6a92f88f1243a6da8164f362cc90b7ebd