Link: Make selectors flexible
The .cdx-mixin-link() mixin assumes that the CSS classes is-red-link and
is-underlined are used for red/underlined links. To allow mixin users to
use different selectors if they want, split the mixin into sub-mixins
that can be used independently. The main mixin still exists and still
works as before, but is now implemented using the sub-mixins.

See the documentation comment in link.less and the usage in LinkDemo.vue
for examples.

Bug: T316895
Change-Id: I72ef166563b629729b1eb08028d06e9ae8040bc1