TextInput: Suppress keydown events for Home/End
This is really aimed at components that combine a TextInput with a Menu,
and delegate keyboard events from the TextInput to the Menu. In that
case, Home/End should not be delegated to the Menu, because we want
Home/End to move the cursor in the TextInput, rather than highlighting
the first/last option in the Menu. Ctrl/Cmd+Home/End are not intercepted,
so that it's still possible to move the Menu highlight with the keyboard.

This logic properly belongs in components that delegate keyboard events
from a TextInput to a Menu, but there are a number of those, and it's
easy to forget to add this logic when we introduce new ones. By having
TextInput lie to its parent and hide Home/End events, we avoid that
maintenance burden.

Bug: T314728
Change-Id: Ib0c0c1c1ad051d6bb4c541e7633ef7221b454ea6