docs: Add more scrolling demos and load-more docs
These features were added to the Menu component and can be used
within components that contain a Menu. This patch adds a demo of
the scrolling behavior to the Combobox and Select pages, and adds
text to the existing fetch demo for Lookup explaining how scroll
is configured and the load-more event is used.

Note that there are no demos of the load-more event for Combobox
or Select because this feature is not likely to be used in those
components.

Bug: T306932
Bug: T316603
Change-Id: I5437ab0a3a8f09a9ecf03cfcea442127526b3f41