tokens: Amend Yellow 600 color option token and add Yellow 500
Amending Yellow 600 to slightly darker `#edab00`.
The former 600 value `#fc3` continues to exist for designs, like
illustrations, hence the addition of Yellow 500 with it as value.
Yellow 600 will be the go-to color for all medium yellow elements like
the warning icon in messages.
Based on Codex Color tokens Figma v23, as of 2022-11-10
https://www.figma.com/file/E6hxGwEOs9YFVF4fFRyqSv/Colors-Tokens---T296995?node-id=1917%3A40831

Bug: T322428
Change-Id: Ic0c1efd0b0d75fec6eed88e56c7bbbefdf9a5b8d