docs, Link: Add page for the Link mixin
This page is temporarily located in the components directory
within the docs so it can quickly be added to the Codex docs site.
Eventually, it will be moved to a more appropriate location, once
we determine the final documentation structure for Less mixins.

Change the alias config in the docs package so that it doesn't break
importing the mixin.

Bug: T317160
Change-Id: I556c9ca22d966e91ff524e4a824aef7996919919