Menu: Remove -1px margin-top from menu
This negative margin was originally intended to tuck the menu's
border-top underneath the input it's attached to. However, the
designs do not include this negative margin, and simply place
the input's borders directly next to the menu's border-top, so
that both are visible

Removing this margin makes the focus outline appear to be the
same on all sides of the input, and fixes a couple of bugs
at different zooms in Chrome.

This change made more visible an existing bug in Combobox, where
the bottom right border radius of the button needs to be set to
0 when expanded so that the button's and menu's borders touch
at a right angle.

Bug: T306932
Change-Id: Ic649c7f9f9bd9749aabed4e748c4cdd070630d90