MenuItem: Reset line-height to @line-height-x-small
The correct value here is @line-height-small, but this will cause
a visual change in TypeaheadSearch that is not acceptable in its
production use in Vector. For now, we'll use this deprecated
line-height value. Eventually, we will increase the size of the
thumbnail and the line height.

Note that the line height was previously (and erroneously) 1.375,
which ends up being 19.25px at a 14px font size (e.g. on desktop
in MediaWiki). It will now be an even 20px in that context.

Bug: T322383
Change-Id: Ie67be53ad665a15d19fe7b87f905ae71410587ac