build, styles: Introduce 'stylelint-declaration-strict-value' plugin
Introducing the new 'stylelint-declaration-strict-value' plugin, which
lints against using values that are not in the allowed list –
Codex design tokens, SFC variables and keywords.
The keyword list is based on our current agreements including the
normalization and reset values.

Note,
- that introducing SFC variables on top is preferred over
  `stylelint-disable` rules as it makes identifying token gaps easier
- that some properties are not in the list, as they are providing
  only functional or resetting values, that are not meant to be themed:
-- `content`, `background-repeat` and `position`
- that `background-image` doesn't seem to get caught, see
  `repeating-conic-gradient()` in TokenDemo

Also
- replacing ToggleSwitch's value that's already available as token,
- replacing `font-weight: bold` with available token in toggle-button
  and CardGroupWithThumbnails
- moving one static value into an SFC variable.

An excerpt of properties going to be tackled in a follow-up:
Add more properties to the list
- `background-position`
- `top`, `right`, `bottom`, `left`
- `width`, `height`
- `margin`, `border` & `padding` longhands
- `animation-iteration-count`
- …
A number of tasks have already been filed for the missing propertes.

Bug: T322388
Change-Id: I553d4a81d7aed80f6147b0491f5bfd2bfa573670