Icon, docs: Move `fill: currentColor` to CSS
The SVG fill can be handled by the CSS `fill` property and it seems to
be a cleaner separation of concerns here.
Also updating Icons overview to explain `currentColor` appropriately.
“CSS syntax is case-insensitive within the ASCII range
(i.e., [a-z] and [A-Z] are equivalent), except for parts that are not
under the control of CSS” – hence going for `currentcolor` to stay in
Wikimedia coding style.

Change-Id: I16a0a9d8142f3fd8d2231e4b5a9f3712381dfae1