tokens: Introduce `font-family` (legacy) tokens
This introduces all `font-family` values available in WikimediaUI Base.
It also brings all font design choice comments to remain in Codex.

Note, that this is featuring one slight value change, for
`font-family-base` now Vector's default `sans-serif` is used and not
the intermediate choice that never made it anywhere widely into
production and got deprecated as well.
Regarding this, VitePress base font family is updated to remain
unchanged in value.

We might reconsider the sub-grouping of font-family values into
- `sans-serif`,
- `serif` and
- `monospace`
in the future, as it is already putting concrete values in otherwise
abstract theming levels. In design sync we've agreed that themers might
overwrite these theme tokens and would have to document accordingly as
simplest way out for now before typography mixins or further refined
theme font family tokens are added.

Bug: T322811
Change-Id: I48ff41ab20107b75f685444f98cd71a73832b520