docs: Fix direction control for configurable Dialog demo
This configurable demo is different, because controls are manually
added to the body of the dialog, rather than using the existing
infrastructure inside the Wrapper component. This patch makes the
direction control work in this demo, and fixes some overlapping
styles that occur when the Radio components have ancestors with
both LTR and RTL dirs.

Bug: T309589
Change-Id: Ie9e5e54ad73b971010d06546f0d1a840e11a69e1