Dialog: Ensure border-box is used for layout
On the Codex documentation site, global CSS styles apply
`box-sizing: border-box` to all elements on the page. In
MediaWiki on the other hand, this is not applied and most
elements default to the classic `content-box` behavior.

The Dialog styles were written with `border-box` behavior
in mind, so we need to ensure that the appropriate box-sizing
behavior is applied to `.cdx-dialog` and all child elements.

Change-Id: Iac2172ecf3236c40bfd6b3a75dbf0a50d17c92b3