docs, styles: Replace non-existent token for nav spacing
In our custom CSS for the VitePress site, we have two styles that
apply to the outline navigation on the right side of pages. These
two styles used to include a 6px value, which was recently replaced
with a `spacing-35` token. That token doesn't exist though,so the
nav items were too close and the active nav item indicator was
not lined up.

This patch replaces that token with the previously used px value.
Ssince this is just for overrides meant to improve existing Vite
styles, using a specific px value seems acceptable.

Change-Id: Iff5dfe9221609f73f94bcafa996e67b3903fff19