docs: Prevent VitePress from treating the link to latest as internal
When serving the docs at https://doc.wikimedia.org/codex/main/,
VitePress thinks that the link to
https://doc.wikimedia.org/codex/latest/ is an internal link, and tries
to handle it itself without navigating the page. This fails and causes a
404.

Mark the link with class="vp-raw" to prevent this.

(If this link was written in Markdown, VitePress would correctly detect
that it was external and set target="_blank", which also prevents
internal treatment. But this link is written in Vue not Markdown, and we
don't want to set target="_blank" because we don't want it to open in a
new tab.)

Change-Id: I4bd66ea666b56eecbde8011c5b46ecfc64fbfde6