Dialog: Prevent skins from breaking header styles
In Vector (and likely on other skins as well), global
styles on the page could modify the h2 element inside
the Dialog header, causing content headign styels to
be applied erroneously.

Adding "inherit" rules for font-family and font-weight
will help to prevent this. Also, border is zeroed-out
entirely (use the built-in dialog divider elements
instead if a separator line is desired).

Bug: T324495
Change-Id: I5ff733fcde60b5b4b31ece526ff3dce495bc8e47