build: Ensure build-demos script resolves token paths correctly
5b397293c86b61c3bd34d71f378956338ef916a9 added a new alias to the
base Vite config file for Codex so that 'dist' directory “free” file paths for token
imports still get resolved correctly.

That change also needs to be made to the Vite config file in codex-docs,
which is used when the build-demos script is run. This is only used
to embed Codex demos in MediaWiki (the docs site has its own build
process).

Change-Id: Ie5b7fe904f0d25d27af6f331860219a650555bf7