build: Build legacy versions of the Codex styles
This builds codex.style-legacy.css and codex.style-legacy-rtl.css (and
the equivalents for the codex-search package), which use the "legacy"
theme from the tokens package (using a 14px base font size) rather than
the default theme (which uses a 16px base font size).

This is a quick and dirty solution that works by aliasing the
theme-wikimedia-ui.less import to theme-wikimedia-ui-legacy.less in
legacy mode. We should reconsider this when we build a proper theme
system in the future.

Bug: T324367
Change-Id: I5b5c7345dfa68058b32a7b4e47c6ab88b86a48b2