Combobox, Lookup: Add aria-controls attribute
This should be set to the same ID as the aria-owns attribute: the input
is not only a functional parent of the menu (aria-owns), but also
controls its contents and presence (aria-controls).

This was originally pointed out for new-lexeme-special-page’s use of
Wikit by axe-core 4.6 (T326681), but should apply to Codex as well.

Change-Id: Ia9207a6af56e4080c9e8db23553f12edfaa75b2f