docs: Move SVGO preset JSON file to the public/ directory
This reverts commit 8e1ca4ba6216abe2d3156f65582a371adb535c73.

In theory, we should be able to put this file in the public/ directory
and link it using an absolute path like /SVG-foo.json , but that doesn't
work when a docroot is used. This is a bug in VitePress, which I've
filed as https://github.com/vuejs/vitepress/issues/1806

In the meantime, using a relative path works around this issue.

Change-Id: I86367455eea49766090c5cdfe74f8da6b2514731