build: Use shared Vite config for the demos build
This has the effect of building separate LTR and RTL stylesheets for the
demos build in the codex-docs package, as well as separate legacy and
non-legacy builds.

This demos build is only used in the VueTest extension.

Also run the demos build as part of the overall build for the docs
package, and remove the special "npm run build-demos" command in the
docs package. VueTest can use "npm run doc" at the top level instead.
This also builds the tokens and icons packages, without which the build
doesn't work.

Bug: T325560
Change-Id: Ic3448ce8ffef47a1ba2b502cdabb4c100cfc1adb